Made Vector2D and Plane objects, and Graphed them
Vector2D, object that behaves like a vector. It can be added to the right
by 2D arrays. Cannot yet be added to other vectors. Has coordinates and color
attritbutes.

Plane, object that vectors are graphed on. Only has limits attritbutes to show
what coordinates are shown initially.

Graphing has Place_Vector2D function that will place a vector on a plane and
Graph function for showing the plane with the vectors on it.