Can now add Vector2D objects together.
Added new _components argument to Vector2D initialization, which is
a list of other vectors. When two Vector2D objects are added, their
components will be appended to the list of components of the sum.

Added new Place method to Graphing. Has same functionality as
PlaceVector2D, except when it places a sum of vectors, it will
place its components as well to better visualize vector addition.
There is an option to turn this functionality off.